Added Traffic
There are now other cars on the road--two versions of a hatchback, a motorcycle, and a van.  The number of vehicles, as well as their range in speeds, can be configured in the RoadControl script.  New car models can be designed and added to the potential traffic pool by creating new ScriptableObject assets of the class, CarModel.